+++ common-db/src/main/java/cn/featherfly/common/db/dialect/Dialects.java
@@ -2,10 +2,7 @@
 package cn.featherfly.common.db.dialect;
 
 /**
- * <p>
- * Dialects
- * </p>
- * .
+ * Dialects.
  *
  * @author zhongj
  */