+++ common-db/src/main/java/cn/featherfly/common/db/dialect/Keywords.java
@@ -198,6 +198,8 @@ public enum Keywords {
     SCHEMA,
 
     /** The unique. */
-    UNIQUE;
+    UNIQUE,
 
+    /** The collate. */
+    COLLATE;
 }